Change model JSON format
The new format supports having multiple score evaluations for the same
benchmark on the same model, which sometimes occurs when separate
entities evaluate the model with different harnesses or with divergent
findings. It is intended to be easier to edit, and to allow adding
metadata such as the source of the benchmark claim.

We update the score prediction code to use the new format.
We simplify it a bit along the way.
When multiple scores are given, we take the average,
making sure that the source information is preserved.

We also write out the predicted output into a new data file.
That file will eventually be used as a data source for the website.