Benchmark evaluating predictors
We evaluate the quality of the predictors
in a fashion that they cannot game.
We provide to them the raw benchmarks,
with one arbitrary benchmark omitted.
We then look at the difference between
their prediction of the removed benchmark, and the real score.
We do that over many distinct omitted benchmarks,
and compute the mean squared error.

Each benchmark has a widely different range of acceptable values.
One may be between 25 and 100 for instance,
while another might have values between 0 and 4000.
Summing the squared errors across benchmarks is thus ill-defined.
It would be dominated by the error of high-range benchmarks.

Instead, we normalize all scores before computing the MSE.
The normalization yields (score - mean) / standard deviation,
where the mean and standard deviation are taken from the benchmark.

We compare our current score predictor iterated a thousand times,
with a trivial predictor which merely replaces missing values
by the mean of their benchmark:

| Predictor               | Iterations | MSE   | Standard deviations | Latency |
|-------------------------|------------|-------|---------------------|---------|
| Means                   |          1 | 1.108 | 1.053               |  0.095s |
| Multivariate regression |          1 | 0.993 | 0.996               |  0.160s |
| Multivariate regression |         10 | 0.975 | 0.988               |  0.346s |
| Multivariate regression |        100 | 0.873 | 0.934               |  1.985s |
| Multivariate regression |       1000 | 0.893 | 0.945               | 17.986s |