Update gradient descent method to new model format
For the purpose of comparing approaches,
we freshen up the gradient descent approach
so that it works with the current model data loader.

We obtain the following scores:

| Predictor               | Iterations | MSE   | Standard deviations | Latency |
|-------------------------|------------|-------|---------------------|---------|
| Cons. Bivariate reg.    |          1 | 3.031 | 1.741               |  0.222s |
| Cons. Biv. Grad. Desc.  |       1000 | 3.030 | 1.741               | 19.165s |
| Cons. Biv. Grad. Desc.  |      10000 | 3.024 | 1.739               |187.975s |
| Means                   |          1 | 1.122 | 1.059               |  0.195s |
| Multivariate Regression |          1 | 1.046 | 1.023               |  0.272s |
| Multivariate Regression |         10 | 1.026 | 1.013               |  0.546s |
| Multivariate Regression |        100 | 0.899 | 0.948               |  2.602s |
| Multivariate Regression |       1000 | 0.886 | 0.941               | 21.932s |
| Multivariate Regression |      10000 | 0.874 | 0.935               |221.142s |

Keep in mind that the gradient descent approach
relies on formulae associated with what we call
the consolidated bivariate regression,
of which this study highlights the flaws.
It is likely that a gradient descent approach on top of
a better formula, such as the multivariate regression,
would yield much better results.