Add the weighed bivariate regression prediction
An issue with the multivarate regression is of course that
no models have scores across all benchmarks,
and imputting the mean is a poor approximation.
There is no mathematical proof that iterating through
yields increasingly precise results, and empirically, we see
that, on a larger set of model data, some values are stuck, yet
cursory manual analysis can guess those scores better.
Some runs even show a worsening of the precision of estimations
past a given number of iterations.

We explore another approach to predicting values.
It computes the regression coefficients
only between pairs of benchmarks, and averages the predictions
of other benchmarks for its own score estimation.
Stupendously, the quality of this fast algorithm is competitive
with an iterative mutivariate regression that is 50x more costly:

| Predictor               | Iterations | MSE   | Standard deviations | Latency |
|-------------------------|------------|-------|---------------------|---------|
| Cons. Bivariate reg.    |          1 | 3.031 | 1.741               |  0.222s |
| Cons. Biv. Grad. Desc.  |       1000 | 3.030 | 1.741               | 19.165s |
| Cons. Biv. Grad. Desc.  |      10000 | 3.024 | 1.739               |187.975s |
| Weighed Bivariate reg.  |          1 | 0.932 | 0.965               |  0.232s |
| Means                   |          1 | 1.122 | 1.059               |  0.195s |
| Multivariate Regression |          1 | 1.046 | 1.023               |  0.272s |
| Multivariate Regression |         10 | 1.026 | 1.013               |  0.546s |
| Multivariate Regression |        100 | 0.899 | 0.948               |  2.602s |
| Multivariate Regression |       1000 | 0.886 | 0.941               | 21.932s |
| Multivariate Regression |      10000 | 0.874 | 0.935               |221.142s |