Improve the variance estimate for the weighed bivariate regression
We now use the full, correct formula for the bivariate estimation
variance, which takes into account not just the error term,
but now also the uncertainty of the alpha and beta parameters'
estimation.

On top of this, we better estimate the overall weighed average
of the scores that yield the target score, by including
the covariance in the calculation of the final variance.