Update score estimates through gradient descent
After the addition of a larger benchmark dataset,
the loss goes up during training.
The cause is the updatePredictedScores() function:
it looks like the weighed bivariate regression gives such good results,
that even a single execution of the multivariate regression
yields worse results at any step size of gradient descent.

One bug was that the loss worsening detection was on
parameters that were updated in the loop.
We now have a copy of all parameters in the computational graph,
on which we check for the change in loss,
rolling back safely if we see a worse loss, and
changing the step size.

Additionally, we now no longer set the scores to the output of the score estimation
function (which uses the alpha and beta parameters learnt through
gradient descent), instead using the scores learned through
graident descent themselves as well.

While the loss now goes down consistently,
it does look like the benchmark shows worse numbers.