Add metric selection to website
The current website features a sorting system, which lets you select
benchmarks that you want to sort by using a weighed multi-score
criteria. However, if you have multiple sorting formulae (say, one for
coding, and one for math), you need to rebuild it every time you want to
switch between them.

It also displays a table with a huge number of columns, which is
unbearable to navigate. You might be interested in a couple of them,
which might be quite far apart.

We replace the sorting system, with a system to create metrics.
Each metric, when selected, performs the sorting as before,
but it is persistent, and you can come back to it after selecting
another metric.
You also only see the benchmarks associated with the metric.