Use a PRNG for selecting scores in benchmark
The benchmark was selecting scores whose prediction to assess,
based on their name: the benchmark and model names closest to the letter
A were assessed only.

We believe this created a bias, where those benchmarks were hardest to
predict.

We instead now rely on a PRNG to select pseudo-random scores.