Add score estimate to predictors in bivariate variance estimation
We noticed a number of goofy predictions.
Analysis showed that it was caused by a benchmark A which had a number
of scores, but none for which there was a score in benchmark B.
Thus the MSE of the prediction of A by B was 0,
even though it was obviously very very high variance.

We fix it by instead looking at the MSE of predicted scores for B
and seeing how they predict A.