Use infinite variance when there is no bivariate predictability
We build upon the previous commit, by taking care of cases where we
don't even have any pairs of two known benchmark scores to hold onto.
In those cases, there is effectively no way to predict one score from
the other, thus we rely on an infinite variance.

Together with the previous fix, we get a much better prediction
using our benchmark, using the recent improvements to the benchmark
(including a PRNG and more models in the data):

| Algorithm | MSE       |
|-----------|-----------|
| Before    | 38.574831 |
| After     |  0.467999 |