Set capabilities on models and benchmarks
We now set the score of models to no better than random,
when the model does not support the capabilities required
by the benchmark for it to run.

For instance, a benchmark that requires viewing images,
ran on a model which has no way to view those images
because it does not have a vision encoder,
will have the worst score typically achievable on that benchmark.